requirements.txt
streamlit
groq
python-dotenv
pandas
openpyxl
langchain
langchain-groq
langchain-community
sentence-transformers
faiss-cpu